requirements.txt
flask
flask_cors
IPython
python-dotenv